// src/app/models/member-detail-dto.ts
export class MemberDetailDto {
    memberId: number;
    name: string;
    surname: string;
    memberNumber: string;
    tcnumber: string;
    statusParameterId: number;
    associationBranchParameterId: number;
    memberDate: string;
    birthOfDate: string;
    
    birthCityId: number;
    decisionNumber: string;
    decisionDate: string;
    isDisabled: number;
    phoneNumber: string;
    
    motherName: string;
    fatherName: string;
    reasonForExit: string;
    exitDecisionNumber: string;
    exitDate: string;
}